Update after comments and running scalafmt
