+++ app/services/SubmittedProcessService.scala
@@ -16,7 +16,7 @@
 
 package services
 
-import connectors.SubmittedProcessConnector
+import connectors.GuidanceConnector
 import javax.inject.{Inject, Singleton}
 import models.{RequestOutcome, SaveSubmittedProcessResponse}
 import play.api.libs.json.JsValue
@@ -25,10 +25,9 @@ import uk.gov.hmrc.http.HeaderCarrier
 import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
-class SubmittedProcessService @Inject()(connector: SubmittedProcessConnector) {
+class SubmittedProcessService @Inject() (connector: GuidanceConnector) {
 
-  def saveForApproval(process: JsValue)
-                     (implicit ec: ExecutionContext, hc: HeaderCarrier): Future[RequestOutcome[SaveSubmittedProcessResponse]] = {
+  def saveForApproval(process: JsValue)(implicit ec: ExecutionContext, hc: HeaderCarrier): Future[RequestOutcome[SaveSubmittedProcessResponse]] = {
     connector.makeAvailableForApproval(process)
   }
