--- test/mocks/MockSubmittedProcessConnector.scala
@@ -1,43 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package mocks
-
-import connectors.SubmittedProcessConnector
-import models.{RequestOutcome, SaveSubmittedProcessResponse}
-import org.scalamock.handlers.CallHandler
-import org.scalamock.scalatest.MockFactory
-import play.api.libs.json.JsValue
-import uk.gov.hmrc.http.HeaderCarrier
-
-import scala.concurrent.{ExecutionContext, Future}
-
-trait MockSubmittedProcessConnector extends MockFactory {
-
-  val mockConnector: SubmittedProcessConnector = mock[SubmittedProcessConnector]
-
-  object MockSubmittedProcessConnector {
-
-    def makeAvailableForApproval(process: JsValue): CallHandler[Future[RequestOutcome[SaveSubmittedProcessResponse]]] = {
-
-      (mockConnector
-        .makeAvailableForApproval(_: JsValue)(_: ExecutionContext, _: HeaderCarrier))
-        .expects(process, *, *)
-    }
-
-  }
-
-}
